[add pci Optimize file]

+++ DT_Liaoning/src/main/scala/com/dtmobile/spark/biz/gridanalyse/PCIOptimize.scala
@@ -47,9 +47,9 @@ class PCIOptimize(ANALY_DATE: String, ANALY_HOUR: String, SDB: String, DDB: Stri
   def analyse(implicit sparkSession: SparkSession): Unit= {
     import sparkSession.sql
     val t = sql("select operator,value from ltepci_degree_condition where field = 'cellrsrpcoverth'").collectAsList()
-    val sRsrpLap = t.get(0).getString(0)+" "+t.get(0).getDouble(1)
+    val sRsrpLap = t.get(0).getString(0)+" "+t.get(0).getDecimal(1)
     val t2 = sql("select operator,value from ltepci_degree_condition where field = 'adjcellrsrpeffectiveth'").collectAsList()
-    val adjRsrpthresh = t.get(0).getString(0)+" "+t.get(0).getDouble(1)
+    val adjRsrpthresh = t.get(0).getString(0)+" "+t.get(0).getDecimal(1)
     sql("use morpho")
     sql(s"""alter table LTE_MRO_DISTURB_PRETREATE60 add if not exists partition(dt=$ANALY_DATE,h=$ANALY_HOUR)""".stripMargin)
     val selectSql=s"""select t.startTime, t.endTime, t.timeseq,
